+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Added functionality:
   * Because the NextSeq2000 software does not generate a sample sheet that includes a project column, it needs to be created manually.
   * Please see the documentation for the `nextseq2k_nml` for details
 * Added a warning if the base_url does not end in /api/
+* Added GUI option to import config settings from file
+  * This writes the settings loaded to the default config file s.t. the settings persist for automated uploads
 
 Developer changes:
 * Refactored `core/cli_entry.py`